+++ lib/src/dependency.g.dart
@@ -35,20 +35,28 @@ GitDependency _$GitDependencyFromJson(Map json) {
 
 HostedDependency _$HostedDependencyFromJson(Map json) {
   return $checkedNew('HostedDependency', json, () {
+    $checkKeys(json,
+        allowedKeys: const ['version', 'hosted'],
+        disallowNullValues: const ['hosted']);
     var val = new HostedDependency(
         version: $checkedConvert(json, 'version',
             (v) => v == null ? null : _constraintFromString(v as String)),
         hosted: $checkedConvert(json, 'hosted',
-            (v) => v == null ? null : new HostedDetails.fromJson(v as Map)));
+            (v) => v == null ? null : new HostedDetails.fromJson(v)));
     return val;
   });
 }
 
 HostedDetails _$HostedDetailsFromJson(Map json) {
   return $checkedNew('HostedDetails', json, () {
+    $checkKeys(json,
+        allowedKeys: const ['name', 'url'],
+        requiredKeys: const ['name'],
+        disallowNullValues: const ['name', 'url']);
     var val = new HostedDetails(
         $checkedConvert(json, 'name', (v) => v as String),
-        $checkedConvert(json, 'url', (v) => _parseUri(v as String)));
+        $checkedConvert(
+            json, 'url', (v) => v == null ? null : _parseUri(v as String)));
     return val;
   });
 }