+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1
 
 - Fixed name collision with error type in latest `package:json_annotation`.
+- Improved parsing of hosted dependencies.
 
 ## 0.1.0